tests: Add test for compose
The new test uses pseudo[1] to emulate some root operations.  pseudo
has not fully coverage of all the functions used by rpm-ostree, so add
a new libshadow that is LD_PRELOAD'ed before pseudo to add missing
pieces.

At the moment the post-process phase is not tested, as it requires
some changes in execle in the pseudo code (apparently it doesn't honor
the chroot yet), the fork emulation in libshadow forces the new
processes to return immediately.

A local yum repository, which contains a dummy .rpm package adding
support for programs needed by the post-process phase, is created as
part of the test.

Since no libc is present, the empty.c program (which is a no-op)
directly invokes the interrupt 0x80 to exit.

The test works only on x86_64, it is skipped on other architectures.

1) https://www.yoctoproject.org/tools-resources/projects/pseudo

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>